requirements.txt
openai
requests
pyyaml